Merges fuzztool into envtool  (#2645)
